+++ core/amber/src/main/scala/texera/common/tuple/schema/AttributeType.java
@@ -1,4 +1,4 @@
-package Engine.Common.tuple.texera.schema;
+package texera.common.tuple.schema;
 
 import com.fasterxml.jackson.annotation.JsonValue;
